+++ package.json
@@ -5,7 +5,7 @@
     "start": "node .",
     "dev": "ts-node src/index.ts",
     "build": "tsc",
-    "watch": "concurrently \"nodemon -e ts -w ./src -x yarn dev\" \"nodemon -e graphqls -w ./src -x yarn generate:schema-types\"",
+    "watch": "concurrently -n \"ts-node,graphqls\" \"nodemon -e ts -w ./src -x yarn dev\" \"nodemon -e graphqls -w ./src -x yarn generate:schema-types\"",
     "lint": "tslint **/*.ts",
     "generate:schema-types": "ts-node scripts/generate-types.ts"
   },
